+++ packages/grid/_modules_/grid/utils/utils.ts
@@ -63,6 +63,20 @@ export function localStorageAvailable() {
   }
 }
 
+export function getBrowserScrollbarSize(doc: Document, element: HTMLElement): number {
+  const scrollDiv = doc.createElement('div');
+  scrollDiv.style.width = '99px';
+  scrollDiv.style.height = '99px';
+  scrollDiv.style.position = 'absolute';
+  scrollDiv.style.overflow = 'scroll';
+  scrollDiv.className = 'scrollDiv';
+  element.appendChild(scrollDiv);
+  const scrollbarSize = scrollDiv.offsetWidth - scrollDiv.clientWidth;
+  element.removeChild(scrollDiv);
+
+  return scrollbarSize;
+}
+
 // Util to make specific interface properties optional
 export type Optional<T, K extends keyof T> = Pick<Partial<T>, K> & Omit<T, K>;
